requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14.0
drf-yasg>=1.21.0
PyYAML>=6.0

# Optional (useful for testing)
pytest>=7.0
pytest-django>=4.5
